move services to designated folders
